feat: :sparkles: Fifth Query Completed
Fifth Query:  Make an EndPoint that allows inserting records in the table of inventories, the input parameters should be (id_product, id_warehouse, quantity). • The table cannot repeat the combination of Warehouse I Product Therefore therefore it will be necessary to validate if the income that is already being made does it exist or is it a totally new combination. • If it is a totally new combination, you must do an lnsert, considering the data entered. • If it is an existing combination, then an Update must be done to this record, considering the sum of the existing quantity with the new amount.  Completed